requirements.txt
azureml-defaults>=1.0.45
glob2
Pillow==9.1.1
tensorflow==2.9.1
numpy==1.23.0
pandas==1.4.3
tensorflow-addons==0.17.1
tqdm==4.64.0
keras==2.9.0
tensorflow-addons==0.17.1